+++ core/docs/changelog.md
@@ -1,6 +1,13 @@
 # TBD
 ### Fixes
 * Update comment in KurtosisBackendType enum
+* Updated to container-engine-lib 0.23.0 which fixes:
+    * Service registration
+    * A bug with getting DockerKurtosisBackend in the API container
+
+### Breaking Changes
+* API container now takes enclaveVolumeSize in megabytes, not gigabytes
+  * Remediation: change all gigabyte-based volume size specs to megabytes
 
 # 1.48.0
 ### Fixes